requirements.txt
fastapi
uvicorn
opencv-python
numpy
ultralytics
PyMySQL
pydantic
email-validator
passlib[bcrypt]
python-jose[cryptography]